Rename jsTransforation to buildTransformation
